+++ angular/lib/src/core/metadata.dart
@@ -13,7 +13,6 @@ export 'metadata/lifecycle_hooks.dart'
         AfterContentChecked,
         AfterViewInit,
         AfterViewChecked,
-        OnChanges,
         OnDestroy,
         OnInit,
         DoCheck;